+++ static/scripts/top_stripe_base.js
@@ -1,14 +1,3 @@
-function get_iso_ts(daysAgo) {
-  const d = new Date();
-  d.setDate(d.getDate() - daysAgo);
-  const hour_diff = Math.abs(d.getTimezoneOffset()) / 60;
-  d.setHours(d.getHours() + 1);
-
-  let iso_d = d.toISOString();
-  iso_d = iso_d.slice(0, -8);
-  return iso_d;
-}
-
 document.getElementById("live_view").onclick = function () {
   location.href = `/live`;
 };
@@ -20,3 +9,17 @@ document.getElementById("historical_data").onclick = function () {
 document.getElementById("stats").onclick = function () {
   location.href = `/stats`;
 };
+
+window.addEventListener("DOMContentLoaded", async function () {
+  const url = `/api/logger_status`;
+  const response = await fetch(url);
+  const datapacket = await response.json();
+  indicator = this.document.getElementById("status_indicator");
+  if (datapacket) {
+    indicator.style.color = "#00be29";
+    indicator.style.textShadow = "0px 0px 8px #00be29";
+  } else {
+    indicator.style.color = "#a10000";
+    indicator.style.textShadow = "0px 0px 8px #a10000";
+  }
+});